:recycle: Refactored working directory to include subfolders
